stockのMVCの階層変更に伴う修正
