+++ app/assets/stylesheets/application.scss
@@ -6,11 +6,10 @@
 @import "variables";
 @import "managers/cuisines";
 @import "managers/foodstuffs";
-// @import "users/cuisines";
-@import "users/stocks";
 @import "cuisines";
 @import "favorites";
 @import "genres";
+@import "stocks";
 @import "tags";
 @import "tops";
 @import "todaysmenus";
