+++ app/views/stocks/_search_form_rawmaterial.html.erb
@@ -1,5 +1,5 @@
 <div class="search_rawmaterial col-10 offset-1">
-  <%= form_with(url: search_rawmaterial_users_stocks_path, method: :get, local: true) do |f| %>
+  <%= form_with(url: search_rawmaterial_stocks_path, method: :get, local: true) do |f| %>
     <div class="form-group row">
       <label for="keyword" id="label_keyword" class="col-sm-2 offset-sm-3 col-form-label text-sm-right">食材名</label>
       <div class="col-sm-5">
