+++ app/views/stocks/_search_result.html.erb
