+++ app/views/stocks/create.html.erb
