+++ app/views/stocks/destroy.html.erb
