+++ app/views/stocks/edit.html.erb
