+++ app/views/stocks/index.html.erb
@@ -1,12 +1,14 @@
 <div class="content_wrapper">
   <%= render 'shared/flash_messages' %>
   <div id="heading" class="row my-2 mx-2">
-    <div class="col-12 col-lg-6">
+    <div class="col-10 offset-1 col-lg-6">
       <h1 class="text-center text-lg-left">家にある食材</h1>
     </div>
-    <div id="newStockBtn" class="col-12 col-lg-6">
+    <div id="newStockBtn" class="col-10 offset-1 col-lg-3 offset-lg-3">
       <div class="text-center text-lg-right">
-        <%= link_to '食材を追加する', new_users_stock_path, class: 'btn btn-secondary ' %>
+        <%= link_to new_stock_path do %>
+          <button class="btn btn-outline-secondary btn-block" type="button">食材を追加する</button>
+        <% end %>
       </div>
 
     </div>
