+++ app/views/stocks/search_rawmaterial.js.erb
