+++ app/views/stocks/show.html.erb
