+++ app/views/stocks/update.html.erb
