+++ app/views/users/shared/_all_crud_links.html.erb
@@ -2,7 +2,7 @@
   <div class="col-10 offset-1">
     <%# ストック一覧 %>
     <p class="text-center">
-      <%= link_to 'ストック一覧(stocks/index)',users_stocks_path, class: 'btn btn-success' %>
+      <%= link_to 'ストック一覧(stocks/index)',stocks_path, class: 'btn btn-success' %>
     </p>
   </div>
 <% end %>
