--- spec/requests/users/stocks_request_spec.rb
@@ -1,54 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe "Users::Stocks", type: :request do
-
-  describe "GET /index" do
-    it "returns http success" do
-      get "/users/stocks/index"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-  describe "GET /show" do
-    it "returns http success" do
-      get "/users/stocks/show"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-  describe "GET /new" do
-    it "returns http success" do
-      get "/users/stocks/new"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-  describe "GET /create" do
-    it "returns http success" do
-      get "/users/stocks/create"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-  describe "GET /edit" do
-    it "returns http success" do
-      get "/users/stocks/edit"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-  describe "GET /update" do
-    it "returns http success" do
-      get "/users/stocks/update"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-  describe "GET /destroy" do
-    it "returns http success" do
-      get "/users/stocks/destroy"
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-end
